% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawccplot.R
\name{drawccplot}
\alias{drawccplot}
\title{Exhibit construction of continuity correction when approximating a 
distributoin supported on a lattice by a continuous distribution.
Example is approximation of binomial by normal.}
\usage{
drawccplot(n = 10, z = 4)
}
\arguments{
\item{n}{binomail samle size}

\item{z}{ordinate for which P[Z<=z] is required.}
}
\description{
Exhibit construction of continuity correction when approximating a 
distributoin supported on a lattice by a continuous distribution.
Example is approximation of binomial by normal.
}
\examples{
drawccplot()
}
