% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparedensityplot.R
\name{fun.comparedensityplot}
\alias{fun.comparedensityplot}
\title{Plot various distributions: Cauchy, normal, and uniform.}
\usage{
fun.comparedensityplot()
}
\description{
Plot various distributions: Cauchy, normal, and uniform.
}
\examples{
fun.comparedensityplot()

}
