% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{fun.comparepower}
\alias{fun.comparepower}
\title{Compare power of various one- and two-sample tests via simulation.}
\usage{
fun.comparepower(
  samp1sz = 10,
  samp2sz = NA,
  dist = "rnorm",
  ndim = 1,
  distb = NULL,
  level = 0.05,
  hypoth = 0,
  nsamp = 1000,
  alternative = "two.sided",
  altvalue = NULL,
  flip = F
)
}
\arguments{
\item{samp1sz}{the size of the first sample.}

\item{samp2sz}{the size of the second sample.  Defaults to NA, triggering one-sample tests.}

\item{dist}{vector of names of random number generators, determinating distributions to be compared.  Defaults to just normal.}

\item{ndim}{Dimension of distributions to be compared.  Defaults to one.}

\item{distb}{vector of names of random number generators for second sample.  Defaults to distribution for the first sample.}

\item{level}{level of tests.}

\item{hypoth}{vector null hypothesis values.  For uniform on [0,1] this should be taken as .5.}

\item{nsamp}{Number of Monte Carlo samples.}

\item{alternative}{alternative, of the form aceptable by t.test, z.test, binom.test, and wilcox.test.  Defaults to "two.sided".  Can be a vector.}

\item{altvalue}{vector of alternative hypothesis values.  Defaults to null value for calculation of actual test level.}

\item{flip}{logical indicator to allow asymetric distributions to be flipped.}
}
\value{
A five-dimensional array of test levels (if altevalue equals hypoth) or power otherwise.  Dimensions are test, distribution, alternative, alternative value, and first dimension sample size.
}
\description{
Compare power of various one- and two-sample tests via simulation.
}
\details{
Tests are given by fun.onesampletests or fun.twosampletests, as appropriate.
}
\examples{
fun.comparepower(10,15,nsamp=100)

}
