% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{fun.feiller}
\alias{fun.feiller}
\title{Diagram for the construction of Feiller confidence intervals}
\usage{
fun.feiller(xv, yv, z)
}
\arguments{
\item{xv}{vector of X means}

\item{yv}{vector of Y means}

\item{z}{critical value for Z test.}
}
\description{
Diagram for the construction of Feiller confidence intervals
}
\examples{
fun.feiller(c(1,3,3),c(3,1.5,.3),1.96)
}
