% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{fun.givescore}
\alias{fun.givescore}
\title{Compare normal, van der Waerden, or Savage scores.}
\usage{
fun.givescore(data, sv = NULL)
}
\arguments{
\item{data}{Vector of data to score}

\item{sv}{Type of scores, if character "ns", "vw", or "ss", or vector of scores if numeric.  Defaults to rank scores.}
}
\value{
A vector of scores.
}
\description{
Compare normal, van der Waerden, or Savage scores.
}
\details{
Tests are given by fun.onesampletests or fun.twosampletests, as appropriate.
}
\examples{
fun.givescore(seq(15))

}
