% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{fun.invertsigntest}
\alias{fun.invertsigntest}
\title{Diagram for the inversion of the sign test to get confidence intervals for a quantile, or, in the two-sample case, the Hodges-Lehman estimator.}
\usage{
fun.invertsigntest(
  dataset,
  alpha = 0.05,
  maint = NULL,
  tau = 0.5,
  returndata = FALSE,
  npts = 10000
)
}
\arguments{
\item{dataset}{Data vector, or list of two vectors representing two samples.}

\item{alpha}{1-confidence value; defaults to 0.05}

\item{maint}{title for plot.}

\item{tau}{quantile to be estimated.  Defaults to 0.5.}

\item{returndata}{Logical flag determining whether the data set is included as part of the response.}

\item{npts}{number of points at which to evaluate test prior to inversion.}
}
\description{
Diagram for the inversion of the sign test to get confidence intervals for a quantile, or, in the two-sample case, the Hodges-Lehman estimator.
}
\examples{
fun.invertsigntest(rnorm(12))
}
