% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{fun.onesampletests}
\alias{fun.onesampletests}
\title{Calculate a vector of one-sample tests of the null hypothesis of median 0}
\usage{
fun.onesampletests(x, alternative = "two.sided")
}
\arguments{
\item{x}{the data vector}

\item{alternative}{alternative, of the form aceptable by t.test, z.test, binom.test, and wilcox.test}
}
\value{
A vector of p-values
}
\description{
Calculates z, sign, and wilcoxon rank sum tests of the null hypothesis of center 0
}
\examples{
fun.onesampletests(rnorm(10))

}
