% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{fun.plotexactci}
\alias{fun.plotexactci}
\title{Plot the results of a per-value structure of confidence intervals.  Vestigial.}
\usage{
fun.plotexactci(tauv, cistr)
}
\arguments{
\item{tauv}{Vector of ordinates for the empirical cdf}

\item{cistr}{matrix with two rows, and as many colums as elements in tauv.  First row is lower end of CI, and second row is upper end of CI.}
}
\description{
Plot the results of a per-value structure of confidence intervals.  Vestigial.
}
