% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{fun.studentizedcaucyplot}
\alias{fun.studentizedcaucyplot}
\title{Calculate and plot the density of the studentized cauchy means, via simulation.}
\usage{
fun.studentizedcaucyplot(nobs, nsamp = 10000)
}
\arguments{
\item{nobs}{Sample size for studentized means.}

\item{nsamp}{Number of Monte Carlo samples.  Defaults to 10000.}
}
\value{
A five-dimensional array of test levels (if altevalue equals hypoth) or power otherwise.  Dimensions are test, distribution, alternative, alternative value, and first dimension sample size.
}
\description{
Calculate and plot the density of the studentized cauchy means, via simulation.
}
\examples{
fun.studentizedcaucyplot(10,nsamp=100)

}
