% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testreg.R
\name{fun.testreg}
\alias{fun.testreg}
\title{Compare coverage of confidence intervals from various regression techniques.  Intervals are nominal coverage .1.}
\usage{
fun.testreg(dists = "rnorm", npergp = 10, mcsamp = 10000)
}
\arguments{
\item{dists}{vector of names of random number generators for sample.  Defaults to normal.}

\item{npergp}{Number of observations per sample; defaults to 10}

\item{mcsamp}{Number of Monte Carlo samples.}
}
\value{
A matrix of coverage levels.  Dimensions are distribution and confidence interval technique.

A list with two components.  The first is a two-dimensional array of average interval lengths.  The second is coverage levels.  Dimensions of both are technique and distribution.
}
\description{
Compare coverage of confidence intervals from various regression techniques.  Intervals are nominal coverage .1.
}
\examples{
library(VGAM)
fun.testreg(mcsamp=100,dists=c("rnorm","rcauchy","rlaplace","runif","rexp"))

}
