% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{fun.testregboot}
\alias{fun.testregboot}
\title{Compare coverage of bootstrap confidence intervals from regression.}
\usage{
fun.testregboot(dists, regparam, sampsize = 10, mcsamp = 10000, B = 9999)
}
\arguments{
\item{dists}{vector of names of random number generators for sample.  Defaults to normal.}

\item{regparam}{function to pass to the bootstrapping tool}

\item{sampsize}{Number of observations per sample; defaults to 10}

\item{mcsamp}{Number of Monte Carlo samples.}

\item{B}{Bootstrap size}
}
\value{
A matrix of coverage levels.  Dimensions are distribution and confidence interval technique.
}
\description{
Compare coverage of bootstrap confidence intervals from regression.
}
\examples{
regparam<-function(data,indices,fittedv,residuals){
   y<-fittedv+residuals[indices]
   return(summary(lm(y~data[,1]))$coefficients[2,1:2]^(1:2))}
fun.testregboot(c("rnorm","rcauchy","rexp"),regparam,mcsamp=20,B=99)

}
