% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{hodgeslehmannexample}
\alias{hodgeslehmannexample}
\title{Diagram the construction of the one-sample Hodges-Lehmann estimator}
\usage{
hodgeslehmannexample(zz = NULL, nn = 6, alpha = 0.05)
}
\arguments{
\item{zz}{Data vector.  Defaults to a random Cauchy sample of an appropriate length.}

\item{nn}{Length of data vector.  Ignored if no real vector supplied, and defaults to 6.}

\item{alpha}{1-confidence level.  Defaults to 0.05.}
}
\value{
a vector with two components giving the confidence interval.
}
\description{
Diagram the construction of the one-sample Hodges-Lehmann estimator
}
\examples{
hodgeslehmannexample()
}
