% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertcor.R
\name{invertcor}
\alias{invertcor}
\title{Construct confidence interval for regression slope parameter by inverting correlation tests.}
\usage{
invertcor(xxx, yyy)
}
\arguments{
\item{xxx}{Explanatory vector}

\item{yyy}{Response vector}
}
\value{
A matrix of coverage levels.  Dimensions are distribution and confidence interval technique.

Matrix of confidence intervals
}
\description{
Construct confidence interval for regression slope parameter by inverting correlation tests.
}
\examples{
invertcor(rnorm(10),rnorm(10))

}
