% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertsigntest.R
\name{invertsigntest}
\alias{invertsigntest}
\title{Diagram for the inversion of rank tests to get confidence intervals for a quantile, or, in the two-sample case, location differences.}
\usage{
invertsigntest(
  dataset,
  alpha = 0.05,
  maint = NULL,
  tau = 0.5,
  returndata = FALSE,
  npts = 10000,
  hl = NULL,
  rescaley = FALSE,
  newplot = TRUE
)
}
\arguments{
\item{dataset}{Data vector, or list of two vectors representing two samples.}

\item{alpha}{1-confidence value; defaults to 0.05}

\item{maint}{title for plot.}

\item{tau}{quantile to be estimated.  Defaults to 0.5.}

\item{returndata}{Logical flag determining whether the data set is included as part of the response.}

\item{npts}{number of points at which to evaluate test prior to inversion.}

\item{hl}{logical flag to determine whether to use sign or Mood test (FALSE) or the signed rank or rank sum test (TRUE)}

\item{rescaley}{logical flag to determine whether to rescale the vertical scale to [0,1].}

\item{newplot}{logical flag to start a new plot}
}
\description{
Diagram for the inversion of rank tests to get confidence intervals for a quantile, or, in the two-sample case, location differences.
}
\examples{
invertsigntest(rnorm(12))
}
