% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{location}
\alias{location}
\title{Collection of location measures}
\usage{
location(y)
}
\arguments{
\item{y}{data vector}
}
\value{
mean median, and trimmed mean
}
\description{
Collection of location measures
}
