% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{mmti}
\alias{mmti}
\title{Intermdiate function for calculating Mood median test.  Odd case is calculated recursively. The even-sample size question is reduced either to Fisher's exact test, or to the Gaussian approximation.}
\usage{
mmti(u, rt, ct, exact = FALSE, alternative = "two.sided")
}
\arguments{
\item{u}{One corner of 2x2 contingency table}

\item{rt}{Row totals}

\item{ct}{Column totals}

\item{exact}{Logical determining whether Fisher's test or the approximate counterpart is used.}

\item{alternative}{alternative, defaults to "two.sided".}
}
\value{
A list with one component, the p-value.
}
\description{
Intermdiate function for calculating Mood median test.  Odd case is calculated recursively. The even-sample size question is reduced either to Fisher's exact test, or to the Gaussian approximation.
}
