% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{mood.median.test}
\alias{mood.median.test}
\title{Perform Mood's median test}
\usage{
mood.median.test(x, y, exact = F, alternative = "two.sided")
}
\arguments{
\item{x}{First sample}

\item{y}{Second sample}

\item{exact}{logical, determing whether the exact version is done.}

\item{alternative}{Determines whether two-sided test, or which one-sided test, is done.}
}
\description{
Perform Mood's median test
}
