% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{ranktests}
\alias{ranktests}
\title{Perform a rank-based one- or two-sample test using a Gaussian approximation.}
\usage{
ranktests(x, y = NULL, sv = NULL)
}
\arguments{
\item{x}{First sample}

\item{y}{Second sample.  If default value of NULL, indicates one-sample test.}

\item{sv}{Type of scores, if character "ns", "vw", or "ss", or vector of scores if numeric.  Defaults to rank scores.}
}
\value{
A list of class htest
}
\description{
Perform a rank-based one- or two-sample test using a Gaussian approximation.
}
\examples{
ranktests(rnorm(10),rnorm(15),"ss")

}
