% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reyniplot.R
\name{reyniplot}
\alias{reyniplot}
\title{Draw sample paths associated with the Renyi test based on the log rank statistic.}
\usage{
reyniplot(time = NULL, delta = NULL, group = NULL, nsamp = 100, p = 1)
}
\arguments{
\item{time}{event/censoring times.  If null, hypothetical data set will be censored and plotted.}

\item{delta}{censoring indicator}

\item{group}{variable defining group.}

\item{nsamp}{If time is null, number of random observations.}

\item{p}{if data are random, gives multiplier for alternative.  p=0 for null.}
}
\value{
A list with components m for the statistic, x for ordered times, and y for statistic matrix.
}
\description{
Draw sample paths associated with the Renyi test based on the log rank statistic.
}
\examples{
reyniplot(nsamp=20)

}
