% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{scale}
\alias{scale}
\title{Collection of scale measures}
\usage{
scale(y)
}
\arguments{
\item{y}{data vector}
}
\value{
standard deviation, interquartile range, and mean absolute deviation.
}
\description{
Collection of scale measures
}
