% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showfieller.R
\name{showfieller}
\alias{showfieller}
\title{Diagram the construction of Fieller's confidence interval, for the ratio of expectations of barw and bary.}
\usage{
showfieller(barw, bary, tau, sigma, rho = 0, alpha = 0.05, npts = 60)
}
\arguments{
\item{barw}{Denominator of method of moments estimator.}

\item{bary}{Numerator of method of moments estimator.}

\item{tau}{Standard deviation of Denominator}

\item{sigma}{Standard deviation of Numerator}

\item{rho}{Correlation of numerator and denominator}

\item{alpha}{test level}

\item{npts}{number of points for plot.}
}
\description{
Diagram the construction of Fieller's confidence interval, for the ratio of expectations of barw and bary.
}
\examples{
# Standard case with confidence region as an interval.
showfieller(1, 1, 0.70, 0.70)
# Problematic case with confidence region as complement of interval
showfieller(1, 1, 0.20, 0.20)
# Problematic case with confidence region as entire real line.
showfieller(1, 1, 0.75, 0.75)

}
