% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{showmultigroupscoretest}
\alias{showmultigroupscoretest}
\title{Diagram the critical region for a three-group score test, as a function of the first two dimensions.}
\usage{
showmultigroupscoretest(nrep, scores = NULL, docontour = TRUE)
}
\arguments{
\item{nrep}{Number in each group.}

\item{scores}{Scores.}

\item{docontour}{logical value controlling drawing contour.}
}
\description{
Diagram the critical region for a three-group score test, as a function of the first two dimensions.
}
\examples{
showmultigroupscoretest(c(3,3,4))
}
