% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{siegel.tukey.ranks}
\alias{siegel.tukey.ranks}
\title{Siegel-Tukey rank function.}
\usage{
siegel.tukey.ranks(x)
}
\arguments{
\item{x}{the data vector}
}
\value{
a vector with the same length as the input, with Siegel-Tukey Ranks.
}
\description{
Siegel-Tukey rank function.
}
\details{
The complete test in one place is at https://raw.github.com/talgalili/R-code-snippets/master/siegel.tukey.r
}
\examples{
siegel.tukey.ranks(c(rnorm(10),10*rnorm(10)))
}
