% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squaresplot.R
\name{squaresplot}
\alias{squaresplot}
\title{Draw contours of OLS regression as a function of the two parameters.}
\usage{
squaresplot(x, y, centerx = FALSE, npts = 100)
}
\arguments{
\item{x}{explanatory variable}

\item{y}{response variable}

\item{centerx}{Logical variable determining whether to center covariate}

\item{npts}{number of lattice points on each margin.}
}
\description{
Draw contours of OLS regression as a function of the two parameters.
}
