% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{stdres2resid}
\alias{stdres2resid}
\title{Inverts the relationship giving standardized resituals from raw residuals as given, for example, in lmwork from library MASS .}
\usage{
stdres2resid(stdres, modelfit = NULL, hat = NULL, stddev = NULL)
}
\arguments{
\item{stdres}{Standardized residuals}

\item{modelfit}{Fit of the linear model}

\item{hat}{Vector of hat values}

\item{stddev}{Sample standard deviation of residuals.}
}
\value{
raw residuals
}
\description{
Inverts the relationship giving standardized resituals from raw residuals as given, for example, in lmwork from library MASS .
}
