% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testare.R
\name{testare}
\alias{testare}
\title{Check whether powers are the same for sample sizes in the ratio indicated by asymptotic relative efficiency are similar, for sign and t tests, for Laplace variables.}
\usage{
testare(targetare = 1, nvec = c(20, 100, 1000, 10000), nsamp = 10000)
}
\arguments{
\item{targetare}{Putative relative efficiency, taking a value between 0 and 1.  Set to .5 for Laplace and t to sign test comparison.}

\item{nvec}{vector of larger sample sizes to try.}

\item{nsamp}{number of Monte Carlo samples.}
}
\value{
A table of powers.
}
\description{
Check whether powers are the same for sample sizes in the ratio indicated by asymptotic relative efficiency are similar, for sign and t tests, for Laplace variables.
}
\examples{
testare(.5,nsamp=100)

}
