% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{testjack}
\alias{testjack}
\title{Test bias of the jackknife approach.}
\usage{
testjack(
  fcns,
  dists = rexp,
  nsamp = 1000,
  npersamp = 11,
  nbig = NA,
  others = NULL
)
}
\arguments{
\item{fcns}{functions to jackknife.  Must be given as a list.}

\item{dists}{Random number generator, or list of random number generators.}

\item{nsamp}{number of data sets to sample}

\item{npersamp}{number of observations per sample}

\item{nbig}{number of Monte Carlo examples to get target value.}

\item{others}{list of additional arguments for entries in components of fcns.}
}
\value{
Expectation, jackknife bias, and true value (via large MC)
}
\description{
Test bias of the jackknife approach.
}
\examples{
testjack(list(mean,median,mean),dists=list(rexp,rnorm),
   others=list(NULL,NULL,list(trim=0.25)),nsamp=5)
}
