% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{fun.testboot}
\alias{fun.testboot}
\title{Compare coverage of bootstrap confidence intervals.}
\usage{
fun.testboot(
  fcn,
  sampsize = 10,
  mcsamp = 10000,
  B = 9999,
  dists = c("rnorm", "rcauchy"),
  truem = NA,
  alpha = 0.05
)
}
\arguments{
\item{fcn}{function to be bootstrapped.}

\item{sampsize}{Number of observations per sample; defaults to 10}

\item{mcsamp}{Number of Monte Carlo samples.}

\item{B}{Bootstrap size}

\item{dists}{vector of names of random number generators for sample.  Defaults to normal and Cauchy.  This fails for a single distribution.}

\item{truem}{True parameter value; should be of length 1 or length of distribution vector.}

\item{alpha}{1-confidence level.}
}
\value{
A list with two components.  The first is a two-dimensional array of average interval lengths.  The second is coverage levels.  Dimensions of both are distribution and confidence interval technique.
}
\description{
Compare coverage of bootstrap confidence intervals.
}
\examples{
fun.testboot(function(x,indices) return(c(mean(x[indices]),1)),mcsamp=100)

}
